cleanup some pmd and findbugs warnings
