+++ contextvh/src/test/java/contextvh/translators/ContextJ2BuildingsTest.java
@@ -31,12 +31,12 @@ import static org.mockito.Mockito.verify;
  */
 @RunWith(PowerMockRunner.class)
 @PrepareForTest(Building.class)
-public class J2BuildingsTest {
+public class ContextJ2BuildingsTest {
 
     private Translator translator = Translator.getInstance();
 
     /**
-     * Test whether Java to Building({@link J2Building}) asks for
+     * Test whether Java to Building({@link ContextJ2Building}) asks for
      * the correct properties of the indicator.
      *
      * @throws TranslationException thrown if the translate method fails.
@@ -54,7 +54,7 @@ public class J2BuildingsTest {
 
         J2Category j2c = PowerMockito.spy(new J2Category());
         J2MultiPolygon j2mp = PowerMockito.spy(new J2MultiPolygon());
-        translator.registerJava2ParameterTranslator(new J2Building());
+        translator.registerJava2ParameterTranslator(new ContextJ2Building());
         translator.registerJava2ParameterTranslator(j2c);
         translator.registerJava2ParameterTranslator(j2mp);
